Changed OAuthToken.expiresIn from long to Long.
This patch changes the stored datatype to the object Long, rather
than the type long, and adds convenience setters for other non-long
values.

+++ kangaroo-servlet-oauth2/src/test/java/net/krotscheck/kangaroo/servlet/oauth2/resource/TokenResponseEntityTest.java
@@ -55,7 +55,7 @@ public final class TokenResponseEntityTest {
 
         TokenResponseEntity entity = TokenResponseEntity.factory(token, state);
         Assert.assertEquals(token.getId(), entity.getAccessToken());
-        Assert.assertEquals(token.getExpiresIn(),
+        Assert.assertEquals(token.getExpiresIn().longValue(),
                 (long) entity.getExpiresIn());
         Assert.assertEquals("Bearer", entity.getTokenType().toString());
         Assert.assertEquals("debug test", entity.getScope());
@@ -91,7 +91,7 @@ public final class TokenResponseEntityTest {
         TokenResponseEntity entity = TokenResponseEntity.factory(token,
                 refresh, state);
         Assert.assertEquals(token.getId(), entity.getAccessToken());
-        Assert.assertEquals(token.getExpiresIn(),
+        Assert.assertEquals(token.getExpiresIn().longValue(),
                 (long) entity.getExpiresIn());
         Assert.assertEquals("Bearer", entity.getTokenType().toString());
         Assert.assertEquals("debug test", entity.getScope());
@@ -121,7 +121,7 @@ public final class TokenResponseEntityTest {
         TokenResponseEntity entity = TokenResponseEntity.factory(token,
                 refresh, state);
         Assert.assertEquals(token.getId(), entity.getAccessToken());
-        Assert.assertEquals(token.getExpiresIn(),
+        Assert.assertEquals(token.getExpiresIn().longValue(),
                 (long) entity.getExpiresIn());
         Assert.assertEquals("Bearer", entity.getTokenType().toString());
         Assert.assertNull(entity.getScope());
